COMP 561 - Computational Biology Methods and Research/Assignment 3/code/requirements.txt
tqdm==4.37.0
numpy==1.14.3
biopython==1.74
gffutils==0.9